+++ src/models/comment/comment.service.ts
@@ -1,6 +1,6 @@
 import IComment from '@comment/types/comment.type';
 import CommentPrismaDto from '@comment/dto/commentDto.prisma';
-import ICommentPayload from '@comment/types/createCommentPayload.type';
+import ICommentPayload from '@comment/types/createCommentPayload.args';
 
 export default function TaskService() {
   //* Get all comments
